Import 3.8.27 from old tar file.
Version LPRng-3.8.27 - Wed Apr 21 11:32:41 PDT 2004

  Fixed:  'Missing HOLD_FILE' logic error fixed in lpd_rcvjob.c
   (Noted by: Mark Tamisiea <tamisiea@jilau1.Colorado.EDU>)
   (And another suggestion by: Christian Reiber <chrei@krischan.org>)

  Fixed: configure with-initpath
    --with-initpath=PATH now works correctly
   (Reported by: Jeff Bastian <jmbastia@ti.com>)

  Fixed (well, clarified):
   Errormsg() now checks for a 'null' error string and provides
   a printable version of errno for logging.

  Edited the Scan_queue() function to print error message and removed
   some dead code.
    (Inspired by comments from: Russell Adams <RLAdams@Kelsey-Seybold.com>)

  lpc MOVE could result in endless loop.
    (Problem reported by: Wichert Akkerman <wichert@wiggy.net> to
	  Debian bug list, forwarded by: Craig Small <csmall at : enc.com.au>)

  lpc MOVE will now start printing job in destination, rather than just copying
    and preserving status of last operation.

  Added detailed error message for connection failure.
    (Inspired by comments by: Russell Adams <RLAdams@Kelsey-Seybold.com>)

  Found a really small (1 byte) memory leak in LPD.  Solves the mystery of
    the server dying with malloc failed messages.
    (Information supplied by: Russell Adams <RLAdams@Kelsey-Seybold.com>)

+++ src/common/lpd_rcvjob.c
@@ -8,7 +8,7 @@
  ***************************************************************************/
 
  static char *const _id =
-"$Id: lpd_rcvjob.c,v 1.68 2004/02/24 19:37:33 papowell Exp $";
+"$Id: lpd_rcvjob.c,v 1.71 2004/05/03 20:24:02 papowell Exp $";
 
 
 #include "lp.h"
@@ -358,13 +358,12 @@ int Receive_job( int *sock, char *input )
 		if( filetype == CONTROL_FILE ){
 			DEBUGF(DRECV2)("Receive_job: receiving new control file, old job.info.count %d, old files.count %d",
 				job.info.count, files.count );
-			if( job.info.count ){
+			if( hold_fd > 0 ){
 				/* we received another control file, finish this job up */
 				if( !discarding_large_job ){
 					if( Check_for_missing_files(&job, &files, error, errlen, 0, hold_fd) ){
 						goto error;
 					}
-					Set_str_value(&job.info,INCOMING_TIME,0);
 				} else {
 					SNPRINTF( error, errlen)
 						_("size %0.3fK exceeds %dK"),
@@ -395,14 +394,11 @@ int Receive_job( int *sock, char *input )
 				goto error;
 			}
 			if( files.count ){
-				/* we have datafiles, FOLLOWED by a control file,
-					followed (possibly) by another control file */
-				/* we receive another control file */
+				/* we have datafiles, FOLLOWED by a control file */
 				if( !discarding_large_job ){
 					if( Check_for_missing_files(&job, &files, error, errlen, 0, hold_fd) ){
 						goto error;
 					}
-					Set_str_value(&job.info,INCOMING_TIME,0);
 				} else {
 					SNPRINTF( error, errlen)
 						_("size %0.3fK exceeds %dK"),
@@ -434,27 +430,28 @@ int Receive_job( int *sock, char *input )
 
 	DEBUGF(DRECV2)("Receive_job: eof on transfer, job.info.count %d, files.count %d",
 		job.info.count, files.count );
-	if( !discarding_large_job ){
-		if( Check_for_missing_files(&job, &files, error, errlen, 0, hold_fd) ){
-			goto error;
+	if( hold_fd > 0 ){
+		if( !discarding_large_job ){
+			if( Check_for_missing_files(&job, &files, error, errlen, 0, hold_fd) ){
+				goto error;
+			}
+		} else {
+			SNPRINTF( error, errlen)
+				_("size %0.3fK exceeds %dK"),
+				jobsize/1024, Max_job_size_DYN );
+			Set_str_value(&job.info,ERROR,error);
+			Set_nz_flag_value(&job.info,ERROR_TIME,time(0));
+			Set_str_value(&job.info,INCOMING_TIME,0);
+			error[0] = 0;
+			if( (status = Set_hold_file( &job, 0, hold_fd )) ){
+				SNPRINTF( error,errlen)
+					"Error setting up hold file - %s",
+					Errormsg( errno ) );
+				goto error;
+			}
+			if( Lpq_status_file_DYN ){ unlink(Lpq_status_file_DYN); }
+			discarding_large_job = 0;
 		}
-		Set_str_value(&job.info,INCOMING_TIME,0);
-	} else {
-		SNPRINTF( error, errlen)
-			_("size %0.3fK exceeds %dK"),
-			jobsize/1024, Max_job_size_DYN );
-		Set_str_value(&job.info,ERROR,error);
-		Set_nz_flag_value(&job.info,ERROR_TIME,time(0));
-		Set_str_value(&job.info,INCOMING_TIME,0);
-		error[0] = 0;
-		if( (status = Set_hold_file( &job, 0, hold_fd )) ){
-			SNPRINTF( error,errlen)
-				"Error setting up hold file - %s",
-				Errormsg( errno ) );
-			goto error;
-		}
-		if( Lpq_status_file_DYN ){ unlink(Lpq_status_file_DYN); }
-		discarding_large_job = 0;
 	}
 
  error:
@@ -478,7 +475,6 @@ int Receive_job( int *sock, char *input )
 			Send_job_rw_timeout_DYN, buffer, safestrlen(buffer), 0 );
 		Link_close( Send_query_rw_timeout_DYN, sock );
 	} else {
-		Set_str_value(&job.info,INCOMING_TIME,0);
 		Link_close( Send_query_rw_timeout_DYN, sock );
 		/* update the spool queue */
 		Get_spool_control( Queue_control_file_DYN, &Spool_control );
